Move and add test for HTML to PDF
