+++ tests/scripts/test_csv_to_pdf.py
@@ -4,7 +4,7 @@ from pathlib import Path
 from click.testing import CliRunner, Result
 from pypdf import PdfReader
 
-from BALSAMIC.assets.scripts.csv_to_pdf import csv_to_pdf, get_table_html_page
+from BALSAMIC.assets.scripts.csv_to_pdf import csv_to_pdf
 
 
 def test_csv_to_pdf(
@@ -30,30 +30,3 @@ def test_csv_to_pdf(
     reader: PdfReader = PdfReader(stream=pdf_path)
     pdf_page: str = reader.pages[0].extract_text()
     assert purity_csv_path.stem in pdf_page
-
-
-def test_get_table_html_page():
-    """Test table insertion in an HTML page"""
-
-    # GIVEN an HTML table and a test table name
-    html_table: str = """
-        <table>
-          <tr>
-            <th>Header 1</th>
-            <th>Header 2</th>
-          </tr>
-          <tr>
-            <td>Data 1</td>
-            <td>Data 2</td>
-          </tr>
-        </table>
-    """
-    table_name: str = "Test Table"
-
-    # WHEN adding the table to an HTML page
-    html_page: str = get_table_html_page(html_table=html_table, table_name=table_name)
-
-    # THEN the table HTML page should be successfully created
-    assert "<html>" in html_page
-    assert f"<h2>{table_name}</h2>" in html_page
-    assert html_table in html_page
